requirements.txt
# Core dependencies
semgrep>=1.27.0             # SAST scanning tool
requests>=2.31.0            # GitHub API calls
python-dotenv>=1.0.0        # Load environment variables (API keys, config)
openai>=1.10.0              # OpenAI API client for LLM integration
anthropic>=0.18.0           # Anthropic Claude API client

# Local ML Agent dependencies (optional - for CodeBERT vulnerability detection)
# Note: Install torch separately with CPU-only version to save space:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
transformers>=4.36.0        # Hugging Face transformers for CodeBERT model